[TE] Support ScaleFabric SHCA NICs (#3591)
Add USE_SHCA build option for ScaleFabric SHCA NICs, which use
extended 17-bit LIDs instead of standard 16-bit IB LIDs. When enabled,
convert LIDs via shca_17b_types.h helpers and widen lid fields to
uint32_t across Transfer Engine and TENT RDMA paths.

Gate yalantinglibs IBV RPC (YLT_ENABLE_IBV) behind USE_SHCA=OFF so SHCA
builds skip ib_socket-based RDMA RPC and fall back to TCP with a warning.

Adjust dependencies.sh to detect SHCA headers and skip conflicting system
RDMA packages (libibverbs-dev on Debian, rdma-core-devel on RHEL).

+++ mooncake-transfer-engine/tests/rdma_test_peers.h
@@ -63,7 +63,7 @@ class RdmaContextTestPeer {
     }
 
     static void seedAutoGidState(RdmaContext &context, ibv_context *verbs_ctx,
-                                 uint8_t port, uint16_t lid, const ibv_gid &gid,
+                                 uint8_t port, uint32_t lid, const ibv_gid &gid,
                                  int gid_index) {
         context.context_ = verbs_ctx;
         context.port_ = port;
